+++ Magic/CHANGELOG.md
@@ -19,6 +19,7 @@
  - Fixed sandblast
  - Renaming a wand on an anvil no longer organizes its inventory (changed default option to false)
  - Add active spell expression
+ - Add basic support for magic bows
  
 ## 7.2.5